add initial database migration and seed data functionality (#61)

+++ tests/__init__.py
@@ -2,54 +2,29 @@ import os
 
 from flask_testing import TestCase
 
+os.environ["ELIXIR_DSS_ENV"] = "test"
+
 from elixir_dss import app, db
-from elixir_dss.models.security import Role, User
+from elixir_dss.models.security import User
 from elixir_dss.models.services import assign_role_to_user, register_new_user
-from elixir_dss.models.submission import (
-    ConsentStatus,
-    ContactType,
-    DeIdentificationType,
-    LegalBasisType,
-    SubjectCategory,
-    SubmissionScope,
-)
+from elixir_dss.models.seed_data import seed_init_data
 
 __author__ = "Pinar Alper"
 
 
 class BaseTest(TestCase):
     def create_app(self):
-        os.environ["ELIXIR_DSS_ENV"] = "test"
         app.config.from_object("elixir_dss.settings.TestConfig")
         return app
 
     def setUp(self):
+        if os.environ.get("ELIXIR_DSS_ENV", "") != "test":
+            raise ValueError(
+                "ELIXIR_DSS_ENV environment variable should be set to test for unittests"
+            )
         db.drop_all()
         db.create_all()
-        initial_data = app.config.get("DATA_INIT")
-
-        for contact_type in initial_data["contact_types"]:
-            db.session.add(ContactType(name=contact_type))
-
-        for name_role in initial_data["names_roles"]:
-            db.session.add(Role(name=name_role))
-
-        for deid_type in initial_data["deidentification_type"]:
-            db.session.add(DeIdentificationType(code=deid_type[0], label=deid_type[1]))
-
-        for subj_cat in initial_data["subject_category"]:
-            db.session.add(SubjectCategory(code=subj_cat[0], label=subj_cat[1]))
-
-        for lb_type in initial_data["legal_basis"]:
-            db.session.add(LegalBasisType(code=lb_type[0], label=lb_type[1]))
-
-        for cons_status in initial_data["consent_status"]:
-            db.session.add(ConsentStatus(code=cons_status[0], label=cons_status[1]))
-        #
-        for sub_scope in initial_data["submission_scope"]:
-            db.session.add(SubmissionScope(code=sub_scope[0], label=sub_scope[1]))
-
-        db.session.commit()
+        seed_init_data()
 
     def tearDown(self):
         db.session.remove()
